Un-revert: Use dictionary lookup while resolving assembly references
This change was originally added to address a O(n^2) algorithm that has been noticed in several customer profiles. Unfortunately, this change was reverted due to a belief that it caused a regression in a speedometer test. However, it appears that is not the case as multiple test insertions with this fix didn't indicate such a regression.

This PR simply adds back the original changes from this PR (https://github.com/dotnet/roslyn/pull/66430) and adds a simple cache around the MultiDictionary allocation.