Enable LSP pull diagnostics from Run Code Analysis command
Fixes #65967

Prior to this change, Run Code Analysis command force executed all the analyzers on the selected project/solution, but the reported analyzer diagnostics were not getting reported from LSP pull handlers. This PR makes the following changes to ensure these diagnostics get reported as workspace pull diagnostics:

1. Modify WorkspacePullDiagnosticHandler so that it reports diagnostics when either of the following conditions are met:
   1. Full solution analysis is enabled (it was already doing so prior to this PR)
   2. Project was already force analyzed and has cached diagnostics (it could be via Run Code Analysis command or any other client in future).
2. Update the Workspace and Project diagnostic sources added by the Workspace pull handler to do the following:
   1. If full solution analysis is enabled, compute and report the closed file diagnostics for the latest solution snapshot
   2. Otherwise, return the already cached closed file diagnostics, which may be from a prior snapshot (this matches the non-LSP behavior when Run Code Analysis command is executed)
3. Update the `DiagnosticAnalyzerService.ForceAnalyzeAsync` command that is invoked during Run Code Analysis to keep track of projects that were force analyzed. Additionally, also send a workspace diagnostic refresh request to the LSP client after each project has finished analysis.

With the above changes, Run Code Analysis command executes and reports diagnostics as expected in LSP pull diagnostics mode. There is still a considerable delay (10 sec) in error list getting populated with diagnostics from this command as the workspace diagnostic refresh is triggered at a delay. We will be addressing this in a separate follow-up change, where in we will trigger a new kind of diagnostic refresh request specifically catered to code analysis execution, such that the client does the workspace pull without any delay. We are still unsure if we need to pass in the project context around in this refresh request OR let it be a full refresh, given the server is only going to attempt to return workspace diagnostics from the cache, so a full refresh request may not incur any significant cost.

TBD: Add unit tests for this scenario. I am going to do this in a follow-up PR so that #70040 does not stay blocked on the product changes here.